backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
websockets==12.0  # WebSocket support for real-time communication

# Database / 데이터베이스
sqlalchemy==2.0.23
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.13.1  # Database migrations

# Authentication / 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# For testing / 테스트용
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.21.1